fix: preserve gRPC status and details when intercepting errors
If the middleware that extracts the gRPC status from an error receives
an argument that already contains a gRPC status structure, the
interceptor respects the status and details and overrides the message
only if an end-user message has been added along the stack.

In some operations like connector updates, the gRPC response is defined
before it reaches the middleware.