feat: support reference with `foo["bar"]` syntax (#394)
Because

- Originally, when performing object traversing, we only supported the
child operator in `foo.bar` syntax. However, this approach failed when
dealing with a key containing a space. To address this limitation, we
needed to implement support for the child operator in `foo["bar"]`
syntax. This change ensures that keys like `my var` can function
correctly with `foo["my var"]`.
- Temporal worker might stuck into infinite retry loop if there is
unexpected error.

This commit

- Supports references with the `foo["bar"]` syntax.
- Changes JSONPath library.
- Updates Temporal worker setting to avoid infinite retry.

Note

- The `openapi.go` is very messy, we'll refactor it soon.