refactor: use combination keys to read/write the trigger memory (#470)
Because

- Originally, we used a single key to read/write the trigger memory
to/from Redis. When using the iterator, we needed to duplicate the
component memory N times (N = size of the iteration batch). In the
original implementation, this caused a lot of duplicated storage. For
the duplicated data in the iterator memory, we should be able to use the
same Redis key across the batch to get the data instead of duplicating
the data.

This commit

- Uses combination keyset to read/write the trigger memory. This way, we
can reuse the same data in the iteration instead of duplicating it.