refactor: remove UseStaticModelList workaround (#499)
Because:

- Originally, we had a config called `UseStaticModelList` to speed up
the Instill Model connector. However, with the introduction of the
Instill Model cloud allowing users to upload models, this config is no
longer needed.
- In the Instill Model connector, the component definition is
dynamically generated by sending a request to Instill Model to get the
model list, which incurs extra communication time. To minimize the
impact on our system, we introduce a new parameter called
`useDynamicDef` to indicate whether we want to use the dynamic
definition or not. We will turn it off in the ListPipelines endpoint
since it does not require the dynamic definition.

This commit:

- Removes the `UseStaticModelList` config.
- Introduces the `useDynamicDef` function parameter.