feat: store recipe as YAML (#515)
Because
- To introduce "recipe as code", we will adopt YAML as the recipe
format. YAML offers advanced features that JSON does not, such as the
ability to include comments. We will gradually transition VDP to use
YAML recipes.

This commit

- Recipes are now stored as YAML in the database. However, since the
frontend has not yet adopted this format, we will continue to convert
the recipes to JSON for the HTTP/gRPC API for now. We utilize GORM hooks
to perform the JSON to YAML conversion.
- The datamodel has been simplified to facilitate easier
marshaling/unmarshaling between struct, JSON and YAML.