feat(component, video): implemented fractional fps value - fps as number, and added some debugging flags for ffmpeg (#765)
Because

- The Video Component minimum FPS is 1 (integer), For a use case like
presentation in the video, we do not want to extract 1 frame every
second but 1 frame every 10 seconds to save resources and reduce the
number of images needed to feed into the next component.
- Issue instill-ai/instill-core/issues/1115.

This commit

- Change FPS to accept real number inputs as well i.e. 60, 30, 1, 0.5,
0.2, 0.1 in both the tasks `TASK_SUBSAMPLE_VIDEO` and
`TASK_SUBSAMPLE_VIDEO_FRAMES`.
- Also implements a string parsing function which can handle fractional
inputs like 1/2 and 1/5 but is commented out - because the input `type`
can only be defined as one single type (`number` or `string` and not
both). The function can be used if this issue is resolved in the future.