fix(run): set pipeline run status as failed when component fails (#836)
Because

- When a component fails to execute due to, e.g., an invalid input, the
status in the pipeline run is `Completed`.
- When a pipeline fails in the `PreTriggerActivity` process (load
information to memory, resolve secrets and connections, initialise
templates), the pipeline run doesn't contain the input or snapshot
recipe.

This commit

- Marks the pipeline run and the component run as failed when any of the
jobs in a component execution fails.
- In the future we might want to have a more sophisticated way to handle
errors. Perhaps a component can fail (e.g. another component is run
given that condition as a fallback) and the pipeline execution can still
be considered valid.
- Splits `PreTriggerActivity` in smaller activities and triggers the
upload (recipe / input) activities as soon as the data is ready in
memory.