chore: delay checking the existence of a connection
Since we're going to allow requesters to provide a connection recipe in
the trigger data, having a recipe that references a missing connection
will become a valid scenario. Connection overwrite values will be
provided by batch, so we only check if a connection doesn't exist at the
moment the batch connections are being built.