fix: use requester's connection, not owner's
We were passing the **pipeline owner UID** to fetch the connections
referenced in a pipeline, which means that other namespaces could access
these connections when triggering a shared pipeline.

The expected behaviour is that the connections should be fetched from
the requester's namespace and produce an error if they don't exist
there.