feat(Docling): prefetch model artifacts (#964)
Because

- Some EasyOCR models are needed by Docling to transform PDF to
Markdown. Without them, the first execution of the document component
fails because the output starts by a "Downloading detection model,
please wait..." print.
- This also prevented coverage for the Docling converter.
- The `use-docling` parameter in the document operator is less open to
changes that an enum `converter` selector.

This commit

- Adds the EasyOCR models to the Docker images.
- Corrects the integration test in the CI after [the latest changes in
instill-core
](https://github.com/instill-ai/instill-core/pull/1172).
- Replaces the `use-docling` parameter by `converter`.

The following changes are made on the `Dockerfile`:
- `nobody:nogroup` needs to have a $HOME where the EasyOCR models will
be placed (internally, this engine looks for the models in
`~/.EasyOCR/model`).
- The workdir (`/pipeline-backend`) is owned by `nobody:nogroup` in the
dev image so we can run the coverage action without the root user.